[UPT] rename module and minor fixes
