--- product_customer_code_picking/models/stock_move.py
@@ -1,26 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2013-2017 Agile Business Group sagl
-#     (<http://www.agilebg.com>)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from odoo import api, fields, models
-
-
-class StockMove(models.Model):
-    _inherit = 'stock.move'
-
-    @api.multi
-    def _get_product_customer_code(self):
-        product_supplierinfo_obj = self.env['product.supplierinfo']
-        for move in self.filtered(lambda m: m.picking_id and
-                                  m.picking_id.partner_id and
-                                  m.product_id.supplier_ids):
-            supplier_info = product_supplierinfo_obj.\
-                search([('product_tmpl_id', '=',
-                       move.product_id.product_tmpl_id.id)], limit=1)
-            move.product_customer_code = supplier_info.product_code or ''
-
-    product_customer_code = fields.Char(
-        compute='_get_product_customer_code',
-        string='Product Customer Code',
-        size=64
-    )
