+++ product_supplierinfo_for_customer_picking/i18n/de.po
