+++ product_supplierinfo_for_customer_picking/i18n/it.po
