+++ product_supplierinfo_for_customer_picking/i18n/product_customer_code_picking.pot
@@ -1,0 +1,27 @@
+# Translation of OpenERP Server.
+# This file contains the translation of the following modules:
+# 	* product_customer_code_picking
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: OpenERP Server 7.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2015-02-01 13:54+0000\n"
+"PO-Revision-Date: 2015-02-01 14:55+0100\n"
+"Last-Translator: Rudolf Schnapka <rs@techno-flex.de>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"X-Generator: Poedit 1.5.4\n"
+
+#. module: product_customer_code_picking
+#: model:ir.model.fields,field_description:product_customer_code_picking.field_stock_move_product_customer_code
+#: field:stock.move,product_customer_code:0
+msgid "Product Customer Code"
+msgstr ""
+
+#. module: product_customer_code_picking
+#: model:ir.model,name:product_customer_code_picking.model_stock_move
+msgid "Stock Move"
+msgstr ""
