+++ product_supplierinfo_for_customer_picking/views/stock_picking_view.xml
